media-source/SourceBuffer-mode_removed.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Setting SourceBuffer#mode after the sourceBuffer object is removed</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var video = document.querySelector('video');
var mimes = ['video/webm; codecs="vorbis,vp8"', 'video/mp4'];

//check the browser supports the MIME used in this test
function canPlayType(mime) {
    if(video.canPlayType(mime) != 'probably') {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test.");
        });
        this.done();
        return false;
    }
    return true;
}
function test(mime) {
    async_test(function(t) {
        if(!canPlayType.bind(t)(mime)) {
            return;
        }
        var mediaSource = new MediaSource();
        mediaSource.addEventListener('sourceopen', function(e) {
            var sourceBuffer = mediaSource.addSourceBuffer(mime);
            mediaSource.removeSourceBuffer(sourceBuffer);
            t.step(function() {
                assert_throws({name: 'InvalidStateError'},
                              function() {
                                  sourceBuffer.mode = 'sequence';
                              },
                              'Check for setting mode value of removed sourceBuffer object.');
            });
            t.done();
        });
        video.src = window.URL.createObjectURL(mediaSource);
    }, 'Setting SourceBuffer#mode (' + mime + ') : ' +
       'If sourceBuffer object has been removed from the sourceBuffers attribute of the parent media source, ' +
       'then INVALID_STATE_ERR exception should be thrown.');
}
mimes.forEach(function(mime) {
    test(mime);
});
</script>
</body>
</html>
